feat(tailwind): Switch to daisy
Used something that (at least to me) feels _way_ more
productive and getting up to speed. Since it's tailwind
under the hood, but with daisy with some components,
we strike a balance of accessibility and onboaridng speed
so candidates are faffing about with grommet, a relatively
obscure library. Had to switch modal/datetime picker